requirements.txt
# Audio-Only Drama — Automated FX Engine
# Python dependencies for the backend

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Task queue and caching
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Audio processing
soundfile==0.12.1
librosa==0.10.1
ffmpeg-python==0.2.0
pydub==0.25.1
scipy==1.11.4
numpy==1.24.3

# Progress bars
tqdm==4.66.1

# WhisperX for speech processing
whisperx==3.1.1

# PyTorch (will be installed separately based on GPU availability)
# torch
# torchvision
# torchaudio

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# PDF processing for screenplay parsing
PyPDF2==3.0.1

# Audio alignment and processing
whisperx==3.1.1
librosa==0.10.1
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3

# SFX generation and audio processing
aiohttp==3.9.1
soundfile==0.12.1
scipy==1.11.4

# NLP and text processing
spacy==3.7.2
